Only set STATIC_ROOT if provided
Currently drama-free-django always sets the Django setting STATIC_ROOT.
This change makes it possible to skip this, by setting the "static_out"
path to a falsey value.

This also slightly modifies the behavior of get_path_if_exists to allow
for empty paths, still raising an error if the path key doesn't exist.