Make extra STATICFILES_DIRS optional
Currently drama-free-django always tries to append to the Django
STATICFILES_DIRS setting with a list of subdirectories of two locations,
specified by the 'static_in' and 'build_static_in' keys. These default
to two different locations for a 'static.in' subdirectory.

Currently cf.gov relies on this behavior to pull in staticfiles from a
static.in subdirectory, but it would be nice to do this directly in
Django settings, instead of relying on DFD.

This change makes it possible to disable this behavior by setting one or
both of the 'static_in' and 'build_static_in' keys to None.